remove useless code

+++ Athena/Player/Decoder/SCHardwareDecoder.m
@@ -108,9 +108,9 @@ static void didDecompress(void *decompressionOutputRefCon,
             if(decodeStatus == kVTInvalidSessionErr) {
                 NSLog(@"IOS8VT: Invalid session, reset decoder session");
             } else if(decodeStatus == kVTVideoDecoderBadDataErr) {
-                NSLog(@"IOS8VT: decode failed status=%ld(Bad data)", decodeStatus);
+                NSLog(@"IOS8VT: decode failed status=%d(Bad data)", (int)decodeStatus);
             } else if(decodeStatus != noErr) {
-                NSLog(@"IOS8VT: decode failed status=%ld", decodeStatus);
+                NSLog(@"IOS8VT: decode failed status=%d", (int)decodeStatus);
             }
 //            NSLog(@"Read Nalu size %d", packet.size);
             CFRelease(sampleBuffer);
